backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
chromadb>=0.4.22
python-frontmatter==1.1.0
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.8
sentence-transformers==2.5.1
FlagEmbedding==1.2.8
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
